Fixed the problems I was having last night - turns out it was a slightly larger issue: when I changed the bounds checking for individualsPassed in the *AOE() methods, it turns out some of the individuals were being skipped, specifically skelly4 who was presumed NOT DEAD for one of my tests. I fixed this by having checks to see if the individual was removed from the group, and if so, decrease the number of IndividualsPassed.
So hey, testing revealed a bug. Who'd of thunk it.

Anyway I changed the targeted ability to have a greater range and to hit
skelly who was far away (which is good, shows off the range effect). I
had to modify subsequent rand() calls because the numbers were low, and
consequently not demonstrating the proper range of the abilities.