I had been brainstorming ways to solve a problem the past few weeks: how do I interrupt running code to display some animations for a few seconds? Specifically displaying hit/miss images over individuals after an attack, and keeping the individual around for a few seconds before they get deleted.
 My first attempt was to create a function that had a clock in it and to
allow it to wait for an arbitrary amount of clock ticks,
however this blocked message queues, and consequently triggered the apps
'unresponsive' mode after 5 seconds of ticks. This really wasn't the way
to go. 

The solution I have now, however, works wonderfully and has room to
grow. The fundamental difference is A) the calculations are not
interrupted, so the attack/damage/death logic is kept in place; however
a static instance of a struct ('specialDraw') contains a list of
characters and individuals, which are drawn for 'x' timer ticks. So, if
an individual is attacked, hit, and killed, after the calculations are
finished the individual is stored in the static instance along with a
'hit' image, drawn on the screen for x ticks, and then removed.

The nice thing about this method is it opens the door for having some
standard animations; I'm thinking that eventually there will be
'animation' structs that are stored in the 'specialDraw' instance, and
will update over timer ticks. For now though, its purpose is well
served.

Also I'm adding a few more sound samples, which I will sync up with the
attack stuff.