Thawak! Bash! Schik! The basic physical-attack sound effects are there, and they sound lovely. Also, I fixed the stupid audio bug I was getting. Of course it's related to not cleaning up my mess, the specific mess being the 'source' field on the soundContainers. Once that got fixed, all my sound problems appear to be solved.
Which is quite nice, because I was encountering the strange sound
Heisenbug again with creating the device/context what not because sounds
were being triggered while in testing and the sound instance was not set
up yet. The solution was to just have a static int that said if sound
was enabled and have it be set to disabled when the tests were running.
That's the solution, and it's staying that way.

Also, with the source-fix in place, inturrupts are working as intended,
with at most two sound effects.