Thanks to my math minor, Line of Sight is now a thing! After tending to first aid, enemies will check if there are any enemies (of the enemies) near by. If the target individuals are within the enemies LoS and not obstructed by any obstacles, then they are registered as the targetedIndividual of the enemy, and will (currently) try to attack them.
LoS is calculated by getting all the squares between the individuals,
which is done by determining the slope between them and taking intervals
with an error of 1/4 (meaning, if the line between the individuals
doesn't go through more than 1/4th of a space, it's not included)

Also, Sound's a total pain in the ass. After my computer crashed (for
unknown reasons), it would hang when trying to obtain the sound device.
Solution was to restart... I'll try to keep this in mind with future
issues.