I've started to bring law to the land - at least the infrastructure for keeping track of crimes.
So the way crime works is that if the player does something in LoS of a
guard, their crime is immediately reported. However, if an NPC sees it,
then their crime is listed as an activeCrime - if they leave the area
without confronting the NPC and succeeding, then their crime will be
reported. however, if they threaten/bribe or kill the witness NPC, that
crime is then removed. Finally, if nobody witnesses the crime, then the
player gets off free (or as I have commented, they're a '//smooth
criminal')