Two changes: now when the user commits a crime where an individual is directly involved (e.g. assault and murder), that individual is aware of the players crime. I may add a spot check to give it more variance.
Also I fixed how cloning works for items - when they're added to the
registry, their ID is equal to the greatest ID in the registry plus 1.
Tested and appears to work correctly.

Found an odd bug - when the player enters/leaves the shop, the view
shift is shifted down too far, and more alarmingly (but likely
unrelated) skellies were able to stop on the same space as the player
(it actually happened 3 times, once with the guard annoyingly). Attempts
to replicate while debugging have not been successful thus far.