Made a few minor changes and some specifications. For one, I decided that the best way to handle when the npc attacks first is to really just make them an enemy, as that was one of the intended purposes of the enemies group. With that, their faction and whether or not they focus on the individual is dependant on the event.
Also, now when a friendly unit is searching for targets to attack, there
is a check to see if they became an ally.

Lastly, I found a strange bug with abilityViewLoop - it was relying on
the return value to denote when the player died from a self inflicted
ability - however there was a windows being returned in there as well,
and returning its value. I took the safe route and just moved the field
into the loop to properly kill the player closer to the declaration.