I've added a new special dialog behavior to separate individuals who are attacked to those who witnessed a crime. In the big picture, the difference is new contextual dialog options, e.g. apologizing for attacking them or returning the item you stole from them.
Anyway that's implemented, which was mostly dialog/event configuration
changes, with some changes to the crime event - specifically, if you
attack an npc out of guard LoS (so the guards dont know you're a threat)
then do it again in LoS, the guards will see you as a threat as well.
You can then yield to either hostile npc or guards to stop both groups
from attacking.