For the coming changes to path finding, I've updated the isAlly function to be more accurate and efficient. It does extra checks concerning interactions between guards, npcs, and their disposition to the player/allies of the player. Also, instead of looping through groups to find the possibleAlly, it just checks the currentGroup enum.
The only major change to the pathing algorithm will be to include the
individual all the way down to where it checks if a possible new space
is passable or not; a space being passable will also depend on if there
are (nonAlly) individuals blocking it, and as such will be considered
impassable. This is required on both the 'searching' and 'actual'
pathing parts, as both require the same isPassable criteria concerning
movement.