Some major changes between now and last commit
The most obvious one is now there are two game modes: ActionMode
(traditional) and FreeMode (free movement, etc.). I fortunately didn't
have nearly as much trouble as I had anticipated implementing this;
there is a minutely check for individuals to see if there are
hostilities near by, or if a guard is pursuing the player - if so the
game is in ActionMode, otherwise it's in FreeMode. Still some bugs to
work out, but for the most part it was a straight-forward
implementation.

I also solved the worry some problem of the NPC performance - turns out
the performance is fine! The root of the issue was I was doing every
individual's action back-to-back, and not allowing the WM_TIMER message
to trigger the draw function! The solution that I came up with in the
shower was to throw the performAction method into a loop-type method
that listens for the timer and draws accordingly. Worked first try.

Just to document a little more about how I came to that solution was by
A) properly measuring the performance of some functions, B) testing what
happens when there are multiple individuals (e.g. move enemies to NPCs),
and C) happened that one of the functions I was testing was also being
used in a timer-loop, which is when I came to the realization that that
timings were the same, just listening when to draw was being handled in
one and not the other.