Now when an individual is attacked and they cannot find a targeted enemy (e.g. they're hiding, attacking between walls, etc) they will move to a random location, weighted to be further away from their starting point. They'll wait there for a few turns before returning to their desired location.
The weighted location selection is kinda neat: first, a list of
coordinates are generated in such a way that the closest ones are first
on the list, with the furthest ones at the end of the list. A random
index of the list is chosen that mimics the function y=x^2, meaning
there's a higher probability of a spot towards the end being selected.
The effect is individuals will typically go to further away spots in a
relatively direction-less manor, simulating how one may react when under
a surprise attack.