I implemented functionality to block attacking through spaces, and to show possible attack spaces to the player. The way it works is it does line of sight between two points, and if any spaces between cannot be attacked through (an attribute on space), it will disallow attacking.
The more interesting algorithm is the one I used to generate all
possible attack spaces--it works basically the same way, except that it
generates a list of the perimeter spaces (the maximum reach of the
attack) and starting from the center, does a LoS out to each perimeter
space. If it encounters a space that cannot be attacked through, it will
return the spaces it has found. This ensures that each space between the
center and perimeter is checked, giving reasonable performance [O(n)
(2n*4, specifically) instead of O(n^2)]

Also I found and fixed a bug with how spaces between cords were being
generated; it became evident when I first displayed possible attack
spaces, the x/y coordinates were being flipped when the calculated slope
was with respect to x.

Lastly, I had to update the tests from top to bottom (again) on account
of implementing the attack-through functionality. To make this a bit
less of a burden, I added some stop-gaps by re-initializing the random
seed, so now if I have to go back and change something, only a
subsection will require updates.

Finally - Ive updated the readme for the first time in half a year.