So after doing a little bit of research, I was able to find a way to speed up the AI process significantly!
I started by doing some timings - it took about .25 of a second to
process AI actions, or roughly 16% of screen time, resulting in player
input being eaten. However, I noticed that when I moused over the
application, everything sped up! In principal, I knew that this was
generating a lot of messages on the queue - and looking at my GetMessage
function, I found that it blocks when there isn't a message to process.
I realized the bottleneck was I had been relying on the WM_TIMER calls
to move the processing along too much, and by placing several
'SendMessage' (actually 'PostMessage) functions in key spots, I was able
to re-enter the loop significantly faster by not blocking on GetMessage.
This is an invaluable trick that I can foresee leveraging in the future.