I solved the problem with individuals moving to their own space first; this was done by including a 'dummy' cord (as well as a flag to use it) in the moveNodeMeta, which would be used in the specific case of moving one space while not in action mode. Basically, the user is moved to their destination space, and the dummy node contains their previous space, and does the opposite offset.
I also fixed an interesting bug; basically, an individual could be added
twice to the movingIndividuals group in the event that their action loop
was triggered before they had completed moving. The fallout being once
the old animation finished, it would clear up the moveNodeMeta and crash
when trying to render the individual again. The fix for this was to make
sure that the action check was always longer than the time it takes to
complete a movement animation, which is possible since they both run on
the same timer (WM_TIMER, specifically).