My branach/repo got messed up (somehow), so im committing what I've got now.
Thanks again to The Forger for his tutorial; I was skeptical if my timer
was working correctly, and looking over his notes on it I found that I
shouldn't be using setTimer for full fledged applications, but rather a
timerQueueTimer. I've started implementing it and, save for some
graphical glitches that I have to figure out, it works amazingly. No
slowdowns, no lag.

Anyway to fix some of those issues I'm working on updating the main
WinProc (since it no longer uses a timer) and I'd like to commit (or try
to) before I muck everything up.