Added a few more locks, game runs reliably.
I noticed that when the window is larger, the draw time increases
(relatively) linearly. I experimented optimizing the way the field is
drawn by fusing the image and imageMask together and drawing only that,
however it's unfortunately not that easy. Actually, since all field
drawings are done on a single DC, I cant really optimize it on a
per-image basis (since the whole image is a collection of the smaller
ones). I'll do some more research concerning if I should switch to a
more efficient drawing system (OpenGL?).