Fixing one thing led to fixing several:
There was a bug with when individuals are doing a long range move, but
aren't in action mode yet (e.g. guards persuing the player). The same
individual would be added to the movingIndividuals group twice, causing
a crash. I fixed that by preventing timer-based-action-mode-check until
the movingIndividuals group was empty. This however lead to a race
condition between an individual doing a long range move to their desired
spot, and another individual going one space. To fix that, I added a
spaceReserved flag to the space struct, which gets set on the final
destination space for long moving individuals, and prevents others from
navigating to them.

I have to fix an issue with how returnToDesiredLocation works; the
faction should be set to -1 only when they're moving one space,
otherwise the full movement shouldn't be blocked.

Also, I fixed how the status read out, and some other minor related
issues.