Implemented Mimics.
Chomp Chomp.

I also updated interactables to process events when attacked - this
includes mimics and chests. While initially I wanted to be a bit more
ambitious and check for how much damage was done and then process it
accordingly, I've opted to remove the randomness both because it's more
consistant and technically much easier to implement; basically when you
attack a chest that's not a mimic, it is immediately destroyed. Likewise
if a mimic is attacked, they're worken up with only 80% of their hp, and
the player keeps their actions.

I've only got a few more interactables I'd like to handle before
considering this done - signs, traps, and leavers.