Possibly the fastest turn-around on a design I've made; got rid of all the description images and replaced them with text. Looks much better. The text is stored in the newGame object, and is currently hardcoded in the init method; if I fee the burning desire (I may), I'll work on reading them in from a file.
Also I fixed a save/load issuewith fate tokens (wasn't being written to
the file).

Lastly, I did my first update to the ability create view: depending on
the status type (which is unfortunately a string and not an enum), the
duration/damage edit fields may not be displayed. The first of likely
several updates to how abilities are created.