Fixed the bug with player/ally position swapping; actually changed quite a bit of it. The big issue being that the player has their own movement handling system which isn't compatable with how all other individuals are handled. Anyway after setting it up properly, both player and ally successfully swap positions.
Also I implemented the first instance of dialog confirmation: when the
player tries to load a game while in the in-game menu, or save over an
existing save, they're prompted for a y/n confirmation. Also as a
side-effect of this, I made it so the saveLoadData gets updated with the
re-initialize, so save confirmations are prompted correctly.